Fixes subscription to _deviceConnector.deviceConnectionStateUpdateStream leaking
- Previous code would create a broadcastStream from a normal stream (created by Repeater).
When this is done, the source stream subscription needs to be explicitly closed.
This is usually done by implementing the onCancel callback from the asBroadcastStream method.
This was not happening, so every call to connectedDeviceStream was creating a new subscription that was never closed.
- To fix the problem, this commit uses the Repeater.broadcast constructor which already returns a broadcast stream
and also closes the source stream subscription when needed